requirements.txt
# Training data generation dependencies
python-dotenv>=1.0.0
google-generativeai>=0.8.0
spacy>=3.7.0
tenacity>=8.2.0
tqdm>=4.66.0
pymupdf>=1.24.0
python-pptx>=0.6.21
python-docx>=1.1.0

# Model training dependencies
torch>=2.0.0
transformers>=4.35.0
scipy>=1.11.0
scikit-learn>=1.3.0
